+++ src/gobexport/exporter/__init__.py
@@ -43,15 +43,12 @@ CONFIG_MAPPING = {
     },
     'brk': {
         'kadastralesubjecten': brk.KadastralesubjectenExportConfig,
-        'aantekeningen': brk.AantekeningenExportConfig,
+        'aantekeningenrechten': brk.AantekeningenrechtenExportConfig,
         'zakelijkerechten': brk.ZakelijkerechtenExportConfig,
         'aardzakelijkerechten': brk.AardzakelijkerechtenExportConfig,
-        'brkbag': brk.BrkBagExportConfig,
         'stukdelen': brk.StukdelenExportConfig,
         'kadastraleobjecten': brk.KadastraleobjectenExportConfig,
         'gemeentes': brk.GemeentesExportConfig,
-        'bijpijling': brk.BijpijlingExportConfig,
-        'perceelnummer': brk.PerceelnummerExportConfig,
         'kadastralegemeentecodes': brk.KadastraleGemeentecodesExportConfig,
         'kadastralesecties': brk.KadastralesectiesExportConfig,
     },
