+++ src/gobexport/exporter/config/brk/kadastraleobjecten.py
@@ -2,8 +2,10 @@ from fractions import Fraction
 
 from gobexport.exporter.csv import csv_exporter
 from gobexport.exporter.esri import esri_exporter
-from gobexport.exporter.utils import convert_format
+from gobexport.exporter.utils import convert_format, get_entity_value
 from gobexport.formatter.geometry import format_geometry
+from gobexport.filters.notempty_filter import NotEmptyFilter
+from gobexport.filters.entity_filter import EntityFilter
 from gobexport.exporter.config.brk.utils import brk_filename, format_timestamp
 
 
@@ -327,6 +329,99 @@ class KadastraleobjectenEsriNoSubjectsFormat(KadastraleobjectenEsriFormat):
         }
 
 
+class PerceelnummerEsriFormat:
+
+    def format_rotatie(self, value):
+        assert not None and isinstance(value, (int, float))
+        # Return rotatie with three decimal places
+        return f'{value:.3f}'
+
+    def get_format(self):
+        return {
+            'BRK_KOT_ID': 'identificatie',
+            'GEMEENTE': 'aangeduidDoorGemeente.naam',
+            'KADGEMCODE': 'aangeduidDoorKadastralegemeentecode.[0].broninfo.omschrijving',
+            'KADGEM': 'aangeduidDoorKadastralegemeente.[0].broninfo.omschrijving',
+            'SECTIE': 'aangeduidDoorKadastralesectie.[0].code',
+            'PERCEELNR': 'perceelnummer',
+            'INDEXLTR': 'indexletter',
+            'INDEXNR': 'indexnummer',
+            'ROTATIE': {
+                'condition': 'isempty',
+                'reference': 'perceelnummerRotatie',
+                'falseval': {
+                    'action': 'format',
+                    'formatter': self.format_rotatie,
+                    'value': 'perceelnummerRotatie',
+                },
+                'trueval': {
+                    'action': 'literal',
+                    'value': '0.000',
+                }
+            },
+            'geometrie': {
+                'action': 'format',
+                'formatter': format_geometry,
+                'value': 'plaatscoordinaten'
+            },
+        }
+
+
+class BrkBagCsvFormat:
+
+    def if_vot_relation(self, trueval: str, falseval: str):
+        return {
+            'condition': 'isempty',
+            'reference': 'heeftEenRelatieMetVerblijfsobject.[0].identificatie',
+            'negate': True,
+            'trueval': trueval,
+            'falseval': falseval,
+        }
+
+    def get_format(self):
+        return {
+            'BRK_KOT_ID': 'identificatie',
+            'KOT_AKRKADGEMEENTECODE_CODE': 'aangeduidDoorKadastralegemeentecode.[0].broninfo.code',
+            'KOT_AKRKADGEMEENTECODE_OMS': 'aangeduidDoorKadastralegemeentecode.[0].broninfo.omschrijving',
+            'KOT_SECTIE': 'aangeduidDoorKadastralesectie.[0].code',
+            'KOT_PERCEELNUMMER': 'perceelnummer',
+            'KOT_INDEX_LETTER': 'indexletter',
+            'KOT_INDEX_NUMMER': 'indexnummer',
+            'KOT_STATUS_CODE': 'status',
+            'KOT_MODIFICATION': '',
+            'BAG_VOT_ID': 'heeftEenRelatieMetVerblijfsobject.[0].bronwaarde',
+            'DIVA_VOT_ID': '',
+            'AOT_OPENBARERUIMTENAAM': self.if_vot_relation(
+                trueval='ligtAanOpenbareruimte.naam',
+                falseval='heeftEenRelatieMetVerblijfsobject.[0].broninfo.openbareruimtenaam'
+            ),
+            'AOT_HUISNUMMER': self.if_vot_relation(
+                trueval='heeftHoofdadres.huisnummer',
+                falseval='heeftEenRelatieMetVerblijfsobject.[0].broninfo.huisnummer'
+            ),
+            'AOT_HUISLETTER': self.if_vot_relation(
+                trueval='heeftHoofdadres.huisletter',
+                falseval='heeftEenRelatieMetVerblijfsobject.[0].broninfo.huisletter'
+            ),
+            'AOT_HUISNUMMERTOEVOEGING': self.if_vot_relation(
+                trueval='heeftHoofdadres.huisnummertoevoeging',
+                falseval='heeftEenRelatieMetVerblijfsobject.[0].broninfo.huisnummertoevoeging'
+            ),
+            'AOT_POSTCODE': self.if_vot_relation(
+                trueval='heeftHoofdadres.postcode',
+                falseval='heeftEenRelatieMetVerblijfsobject.[0].broninfo.postcode'
+            ),
+            'AOT_WOONPLAATSNAAM': self.if_vot_relation(
+                trueval='ligtInWoonplaats.naam',
+                falseval='heeftEenRelatieMetVerblijfsobject.[0].broninfo.woonplaatsnaam'
+            ),
+            'BRON_RELATIE': {
+                'action': 'literal',
+                'value': 'BRK'
+            }
+        }
+
+
 def aandeel_sort(a: dict, b: dict):
     """Returns True if a takes preference over b
 
@@ -351,6 +446,8 @@ class KadastraleobjectenExportConfig:
     csv_format = KadastraleobjectenCsvFormat()
     esri_format = KadastraleobjectenEsriFormat()
     esri_format_no_subjects = KadastraleobjectenEsriNoSubjectsFormat()
+    perceelnummer_esri_format = PerceelnummerEsriFormat()
+    brk_bag_format = BrkBagCsvFormat()
 
     gperc_query = '''
 {
@@ -601,6 +698,189 @@ class KadastraleobjectenExportConfig:
 }
 '''
 
+    bijpijling_query = '''
+{
+  brkKadastraleobjecten(indexletter: "G") {
+    edges {
+      node {
+        identificatie
+        aangeduidDoorGemeente {
+          edges {
+            node {
+              naam
+            }
+          }
+        }
+        aangeduidDoorKadastralegemeentecode {
+          edges {
+            node {
+              broninfo
+            }
+          }
+        }
+        aangeduidDoorKadastralegemeente {
+          edges {
+            node {
+              broninfo
+            }
+          }
+        }
+        aangeduidDoorKadastralesectie {
+          edges {
+            node {
+              code
+            }
+          }
+        }
+        perceelnummer
+        indexletter
+        indexnummer
+        bijpijlingGeometrie
+      }
+    }
+  }
+}
+'''
+
+    perceelnummer_query = '''
+{
+  brkKadastraleobjecten(indexletter: "G") {
+    edges {
+      node {
+        identificatie
+        aangeduidDoorGemeente {
+          edges {
+            node {
+              naam
+            }
+          }
+        }
+        aangeduidDoorKadastralegemeentecode {
+          edges {
+            node {
+              broninfo
+            }
+          }
+        }
+        aangeduidDoorKadastralegemeente {
+          edges {
+            node {
+              broninfo
+            }
+          }
+        }
+        aangeduidDoorKadastralesectie {
+          edges {
+            node {
+              code
+            }
+          }
+        }
+        perceelnummer
+        indexletter
+        indexnummer
+        perceelnummerRotatie
+        plaatscoordinaten
+      }
+    }
+  }
+}
+'''
+
+    brk_bag_query = '''
+{
+  brkKadastraleobjecten {
+    edges {
+      node {
+        identificatie
+        aangeduidDoorKadastralegemeente {
+          edges {
+            node {
+              broninfo
+            }
+          }
+        }
+        aangeduidDoorKadastralegemeentecode {
+          edges {
+            node {
+              broninfo
+            }
+          }
+        }
+        aangeduidDoorKadastralesectie {
+          edges {
+            node {
+              code
+            }
+          }
+        }
+        perceelnummer
+        indexletter
+        indexnummer
+        status
+        heeftEenRelatieMetVerblijfsobject {
+          edges {
+            node {
+              identificatie
+              bronwaarde
+              status
+              broninfo
+                heeftHoofdadres {
+                edges {
+                  node {
+                    identificatie
+                    ligtAanOpenbareruimte {
+                      edges {
+                        node {
+                          naam
+                        }
+                      }
+                    }
+                    huisnummer
+                    huisletter
+                    huisnummertoevoeging
+                    postcode
+                    ligtInWoonplaats {
+                      edges {
+                        node {
+                          naam
+                        }
+                      }
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
+'''
+
+    class VotFilter(EntityFilter):
+        """Only include rows with VOT statuses, for VOT's where status is defined:
+        2   Niet gerealiseerd verblijfsobject
+        3   Verblijfsobject in gebruik (niet ingemeten)
+        4   Verblijfsobject in gebruik
+        6   Verblijfsobject buiten gebruik
+
+        """
+        valid_status_codes = [2, 3, 4, 6]
+        vot_identificatie = 'heeftEenRelatieMetVerblijfsobject.[0].identificatie'
+        status_code = 'heeftEenRelatieMetVerblijfsobject.[0].status.code'
+        city = 'heeftEenRelatieMetVerblijfsobject.[0].broninfo.woonplaatsnaam'
+
+        def filter(self, entity: dict):
+            if get_entity_value(entity, self.vot_identificatie):
+                status_code = get_entity_value(entity, self.status_code)
+                return status_code and int(status_code) in self.valid_status_codes
+            elif get_entity_value(entity, self.city):
+                return not get_entity_value(entity, self.city).lower().startswith('amsterdam')
+
+            return True
+
     sort = {
         'invRustOpKadastraalobjectBrkZakelijkerechten'
         '.invVanZakelijkrechtBrkTenaamstellingen'
@@ -682,5 +962,85 @@ class KadastraleobjectenExportConfig:
                 },
             ],
             'query': esri_query,
+        },
+        'shape': {
+            'exporter': esri_exporter,
+            'api_type': 'graphql_streaming',
+            'secure': True,
+            'filename': lambda: brk_filename('bijpijling', type='shp', append_date=False),
+            'entity_filters': [
+                NotEmptyFilter('bijpijlingGeometrie'),
+            ],
+            'mime_type': 'application/octet-stream',
+            'format': {
+                'BRK_KOT_ID': 'identificatie',
+                'GEMEENTE': 'aangeduidDoorGemeente.naam',
+                'KADGEMCODE': 'aangeduidDoorKadastralegemeentecode.[0].broninfo.omschrijving',
+                'KADGEM': 'aangeduidDoorKadastralegemeente.[0].broninfo.omschrijving',
+                'SECTIE': 'aangeduidDoorKadastralesectie.[0].code',
+                'PERCEELNR': 'perceelnummer',
+                'INDEXLTR': 'indexletter',
+                'INDEXNR': 'indexnummer',
+                'geometrie': {
+                    'action': 'format',
+                    'formatter': format_geometry,
+                    'value': 'bijpijlingGeometrie'
+                },
+            },
+            'extra_files': [
+                {
+                    'filename': lambda: brk_filename('bijpijling', type='dbf', append_date=False),
+                    'mime_type': 'application/octet-stream'
+                },
+                {
+                    'filename': lambda: brk_filename('bijpijling', type='shx', append_date=False),
+                    'mime_type': 'application/octet-stream'
+                },
+                {
+                    'filename': lambda: brk_filename('bijpijling', type='prj', append_date=False),
+                    'mime_type': 'application/octet-stream'
+                },
+            ],
+            'query': bijpijling_query
+        },
+        'perceel_shape': {
+            'exporter': esri_exporter,
+            'api_type': 'graphql_streaming',
+            'secure': True,
+            'filename': lambda: brk_filename('perceelnummer', type='shp', append_date=False),
+            'entity_filters': [
+                NotEmptyFilter('plaatscoordinaten'),
+            ],
+            'mime_type': 'application/octet-stream',
+            'format': perceelnummer_esri_format.get_format(),
+            'extra_files': [
+                {
+                    'filename': lambda: brk_filename('perceelnummer', type='dbf', append_date=False),
+                    'mime_type': 'application/octet-stream'
+                },
+                {
+                    'filename': lambda: brk_filename('perceelnummer', type='shx', append_date=False),
+                    'mime_type': 'application/octet-stream'
+                },
+                {
+                    'filename': lambda: brk_filename('perceelnummer', type='prj', append_date=False),
+                    'mime_type': 'application/octet-stream'
+                },
+            ],
+            'query': perceelnummer_query
+        },
+        'brk_bag_csv': {
+            'exporter': csv_exporter,
+            'entity_filters': [
+                NotEmptyFilter('heeftEenRelatieMetVerblijfsobject.[0].bronwaarde'),
+                VotFilter(),
+            ],
+            'api_type': 'graphql_streaming',
+            'secure': True,
+            'unfold': True,
+            'query': brk_bag_query,
+            'filename': lambda: brk_filename("BRK_BAG"),
+            'mime_type': 'plain/text',
+            'format': brk_bag_format.get_format(),
         }
     }
