+++ src/tests/exporter/config/brk/test_perceelnummer.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from gobexport.exporter.config.brk.perceelnummer import PerceelnummerEsriFormat
+from gobexport.exporter.config.brk.kadastraleobjecten import PerceelnummerEsriFormat
 
 
 class TestPerceelnummerEsriFormat(TestCase):
